Reorganized code in multiple modules
