+++ sentry_sdk/crons/__init__.py
@@ -1,0 +1,2 @@
+from sentry_sdk.crons.api import capture_checkin  # noqa
+from sentry_sdk.crons.decorator import monitor  # noqa
