+++ tests/test_crons.py
@@ -20,7 +20,9 @@ def _break_world(name):
 def test_decorator(sentry_init):
     sentry_init()
 
-    with mock.patch("sentry_sdk.crons.capture_checkin") as fake_capture_checking:
+    with mock.patch(
+        "sentry_sdk.crons.decorator.capture_checkin"
+    ) as fake_capture_checking:
         result = _hello_world("Grace")
         assert result == "Hello, Grace"
 
@@ -34,14 +36,16 @@ def test_decorator(sentry_init):
         # Check for final checkin
         assert fake_capture_checking.call_args[1]["monitor_slug"] == "abc123"
         assert fake_capture_checking.call_args[1]["status"] == "ok"
-        assert fake_capture_checking.call_args[1]["duration"]
+        assert fake_capture_checking.call_args[1]["duration_ns"]
         assert fake_capture_checking.call_args[1]["check_in_id"]
 
 
 def test_decorator_error(sentry_init):
     sentry_init()
 
-    with mock.patch("sentry_sdk.crons.capture_checkin") as fake_capture_checking:
+    with mock.patch(
+        "sentry_sdk.crons.decorator.capture_checkin"
+    ) as fake_capture_checking:
         with pytest.raises(Exception):
             result = _break_world("Grace")
 
@@ -57,7 +61,7 @@ def test_decorator_error(sentry_init):
         # Check for final checkin
         assert fake_capture_checking.call_args[1]["monitor_slug"] == "def456"
         assert fake_capture_checking.call_args[1]["status"] == "error"
-        assert fake_capture_checking.call_args[1]["duration"]
+        assert fake_capture_checking.call_args[1]["duration_ns"]
         assert fake_capture_checking.call_args[1]["check_in_id"]
 
 
@@ -68,7 +72,7 @@ def test_capture_checkin_simple(sentry_init):
         monitor_slug="abc123",
         check_in_id="112233",
         status=None,
-        duration=None,
+        duration_ns=None,
     )
     assert check_in_id == "112233"
 
@@ -82,7 +86,7 @@ def test_capture_checkin_new_id(sentry_init):
             monitor_slug="abc123",
             check_in_id=None,
             status=None,
-            duration=None,
+            duration_ns=None,
         )
 
         assert check_in_id == "a8098c1af86e11dabd1a00112444be1e"
